requirements.txt
playwright
google-generativeai
python-dotenv
requests
streamlit